Putting UFS implementation of journal behind an interface.
