+++ core/server/common/src/main/java/alluxio/master/AbstractMaster.java
@@ -18,12 +18,13 @@ import alluxio.clock.Clock;
 import alluxio.exception.PreconditionMessage;
 import alluxio.master.journal.AsyncJournalWriter;
 import alluxio.master.journal.Journal;
+import alluxio.master.journal.JournalWriter;
+import alluxio.master.journal.ReadWriteJournal;
 import alluxio.master.journal.JournalInputStream;
 import alluxio.master.journal.JournalOutputStream;
 import alluxio.master.journal.JournalTailer;
 import alluxio.master.journal.JournalTailerThread;
-import alluxio.master.journal.JournalWriter;
-import alluxio.master.journal.ReadWriteJournal;
+import alluxio.master.journal.ufs.ReadWriteUfsJournal;
 import alluxio.proto.journal.Journal.JournalEntry;
 import alluxio.retry.RetryPolicy;
 import alluxio.retry.TimeoutRetry;
@@ -48,7 +49,7 @@ import javax.annotation.concurrent.NotThreadSafe;
 public abstract class AbstractMaster implements Master {
   private static final Logger LOG = LoggerFactory.getLogger(AbstractMaster.class);
 
-  public static final long INVALID_FLUSH_COUNTER = -1;
+  private static final long INVALID_FLUSH_COUNTER = -1;
   private static final long SHUTDOWN_TIMEOUT_MS = 10 * Constants.SECOND_MS;
   private static final long JOURNAL_FLUSH_RETRY_TIMEOUT_MS =
       Configuration.getLong(PropertyKey.MASTER_JOURNAL_FLUSH_TIMEOUT_MS);
@@ -114,28 +115,28 @@ public abstract class AbstractMaster implements Master {
        *
        * Phase 2. Mark all the logs as completed. Since this master is the leader, it is allowed to
        * write the journal, so it can mark the current log as completed. After this step, the
-       * current log file will not exist, and all logs will be complete.
+       * current log will not exist, and all logs will be complete.
        *
        * Phase 3. Reconstruct the state from the journal. This uses the JournalTailer to process all
-       * of the checkpoint and the complete log files. Since all logs are complete, after this step,
+       * of the checkpoint and the complete logs. Since all logs are complete, after this step,
        * the master will reflect the state of all of the journal entries.
        *
-       * Phase 4. Write out the checkpoint file. Since this master is completely up-to-date, it
-       * writes out the checkpoint file. When the checkpoint file is closed, it will then delete the
-       * complete log files.
+       * Phase 4. Write out the checkpoint. Since this master is completely up-to-date, it
+       * writes out the checkpoint. When the checkpoint is closed, it will then delete the
+       * complete logs.
        *
        * Since this method is called before the master RPC server starts serving, there is no
        * concurrent access to the master during these phases.
        */
 
       // Phase 1: Recover from a backup checkpoint if necessary.
-      mJournalWriter.recoverCheckpoint();
+      mJournalWriter.recover();
 
       // Phase 2: Mark all logs as complete, including the current log. After this call, the current
-      // log should not exist, and all the log files will be complete.
-      mJournalWriter.completeAllLogs();
+      // log should not exist, and all the logs will be complete.
+      mJournalWriter.completeLogs();
 
-      // Phase 3: Replay all the state of the checkpoint and the completed log files.
+      // Phase 3: Replay all the state of the checkpoint and the completed logs.
       JournalTailer catchupTailer;
       if (mStandbyJournalTailer != null && mStandbyJournalTailer.getLatestJournalTailer() != null
           && mStandbyJournalTailer.getLatestJournalTailer().isValid()) {
@@ -144,7 +145,7 @@ public abstract class AbstractMaster implements Master {
         LOG.info("{}: finish processing remaining journal entries (standby -> master).",
             getName());
         catchupTailer = mStandbyJournalTailer.getLatestJournalTailer();
-        catchupTailer.processNextJournalLogFiles();
+        catchupTailer.processNextJournalLogs();
       } else {
         // This master has not successfully processed any of the journal, so create a fresh tailer
         // to process the entire journal.
@@ -152,14 +153,14 @@ public abstract class AbstractMaster implements Master {
         if (catchupTailer.checkpointExists()) {
           LOG.info("{}: process entire journal before becoming leader master.", getName());
           catchupTailer.processJournalCheckpoint(true);
-          catchupTailer.processNextJournalLogFiles();
+          catchupTailer.processNextJournalLogs();
         } else {
           LOG.info("{}: journal checkpoint does not exist, nothing to process.", getName());
         }
       }
       long latestSequenceNumber = catchupTailer.getLatestSequenceNumber();
 
-      // Phase 4: initialize the journal and write out the checkpoint file (the state of all
+      // Phase 4: initialize the journal and write out the checkpoint (the state of all
       // completed logs).
       JournalOutputStream checkpointStream =
           mJournalWriter.getCheckpointOutputStream(latestSequenceNumber);
@@ -212,7 +213,7 @@ public abstract class AbstractMaster implements Master {
 
   @Override
   public void transitionToLeader() {
-    mJournal = new ReadWriteJournal(mJournal.getDirectory());
+    mJournal = new ReadWriteUfsJournal(mJournal.getLocation());
   }
 
   /**
@@ -223,7 +224,7 @@ public abstract class AbstractMaster implements Master {
   protected void writeJournalEntry(JournalEntry entry) {
     Preconditions.checkNotNull(mJournalWriter, "Cannot write entry: journal writer is null.");
     try {
-      mJournalWriter.writeEntry(entry);
+      mJournalWriter.write(entry);
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
@@ -235,7 +236,7 @@ public abstract class AbstractMaster implements Master {
   protected void flushJournal() {
     Preconditions.checkNotNull(mJournalWriter, "Cannot flush journal: journal writer is null.");
     try {
-      mJournalWriter.flushEntryStream();
+      mJournalWriter.flush();
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
