+++ core/server/common/src/main/java/alluxio/master/journal/Journal.java
@@ -11,7 +11,7 @@
 
 package alluxio.master.journal;
 
-import alluxio.AlluxioURI;
+import java.net.URL;
 
 import javax.annotation.concurrent.ThreadSafe;
 
@@ -25,78 +25,10 @@ import javax.annotation.concurrent.ThreadSafe;
  * completed entry files are in the "completed/" sub-directory.
  */
 @ThreadSafe
-public abstract class Journal {
-  /** The log number for the first completed log file. */
-  public static final long FIRST_COMPLETED_LOG_NUMBER = 1L;
-  /** The directory for completed log files, relative to the base journal directory. */
-  private static final String COMPLETED_DIRECTORY = "completed/";
-  /** The file extension for the current log file. */
-  private static final String CURRENT_LOG_EXTENSION = ".out";
-  /** The filename of the checkpoint file. */
-  private static final String CHECKPOINT_FILENAME = "checkpoint.data";
-  /** The base of the entry log filenames, without the file extension. */
-  private static final String ENTRY_LOG_FILENAME_BASE = "log";
-  /** The directory where this journal is stored. */
-  private final String mDirectory;
-  /** The formatter for this journal. */
-  private final JournalFormatter mJournalFormatter;
+public interface Journal {
 
   /**
-   * Creates a new instance of {@link Journal}.
-   *
-   * @param directory the base directory for this journal
+   * @return the location for this journal
    */
-  public Journal(String directory) {
-    if (!directory.endsWith(AlluxioURI.SEPARATOR)) {
-      // Ensure directory format.
-      directory += AlluxioURI.SEPARATOR;
-    }
-    mDirectory = directory;
-    mJournalFormatter = JournalFormatter.Factory.create();
-  }
-
-  /**
-   * @return the base directory for this journal
-   */
-  public String getDirectory() {
-    return mDirectory;
-  }
-
-  /**
-   * @return the directory for where the completed log files are stored
-   */
-  public String getCompletedDirectory() {
-    return mDirectory + COMPLETED_DIRECTORY;
-  }
-
-  /**
-   * @return the absolute path for the journal checkpoint file
-   */
-  public String getCheckpointFilePath() {
-    return mDirectory + CHECKPOINT_FILENAME;
-  }
-
-  /**
-   * @return the absolute path for the current log file
-   */
-  public String getCurrentLogFilePath() {
-    return mDirectory + ENTRY_LOG_FILENAME_BASE + CURRENT_LOG_EXTENSION;
-  }
-
-  /**
-   * Returns the completed log filename for a particular log number.
-   *
-   * @param logNumber the log number to get the path for
-   * @return The absolute path of the completed log for a given log number
-   */
-  public String getCompletedLogFilePath(long logNumber) {
-    return getCompletedDirectory() + String.format("%s.%020d", ENTRY_LOG_FILENAME_BASE, logNumber);
-  }
-
-  /**
-   * @return the {@link JournalFormatter} for this journal
-   */
-  public JournalFormatter getJournalFormatter() {
-    return mJournalFormatter;
-  }
+  URL getLocation();
 }
