+++ core/server/common/src/test/java/alluxio/master/journal/AsyncJournalWriterTest.java
@@ -20,25 +20,19 @@ import static org.mockito.Mockito.times;
 import alluxio.Configuration;
 import alluxio.ConfigurationTestUtils;
 import alluxio.PropertyKey;
-import alluxio.master.journal.JournalWriter.EntryOutputStream;
 import alluxio.proto.journal.Journal.JournalEntry;
 
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
 import java.io.IOException;
 
 /**
  * Unit tests for {@link AsyncJournalWriter}.
  */
-@RunWith(PowerMockRunner.class)
-@PrepareForTest({JournalWriter.class, EntryOutputStream.class})
 public class AsyncJournalWriterTest {
 
   private JournalWriter mMockJournalWriter;
@@ -57,8 +51,8 @@ public class AsyncJournalWriterTest {
     }
 
     mMockJournalWriter = PowerMockito.mock(JournalWriter.class);
-    doNothing().when(mMockJournalWriter).writeEntry(any(JournalEntry.class));
-    doNothing().when(mMockJournalWriter).flushEntryStream();
+    doNothing().when(mMockJournalWriter).write(any(JournalEntry.class));
+    doNothing().when(mMockJournalWriter).flush();
     mAsyncJournalWriter = new AsyncJournalWriter(mMockJournalWriter);
   }
 
@@ -81,9 +75,9 @@ public class AsyncJournalWriterTest {
       mAsyncJournalWriter.flush(i);
     }
     if (batchingEnabled) {
-      Mockito.verify(mMockJournalWriter, atLeastOnce()).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, atLeastOnce()).flush();
     } else {
-      Mockito.verify(mMockJournalWriter, times(entries)).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, times(entries)).flush();
     }
   }
 
@@ -114,7 +108,7 @@ public class AsyncJournalWriterTest {
 
     // Start failing journal writes.
     doThrow(new IOException("entry write failed")).when(mMockJournalWriter)
-        .writeEntry(any(JournalEntry.class));
+        .write(any(JournalEntry.class));
 
     // Flushes should fail.
     for (int i = 1; i <= entries; i++) {
@@ -127,16 +121,16 @@ public class AsyncJournalWriterTest {
     }
 
     // Allow journal writes to succeed.
-    doNothing().when(mMockJournalWriter).writeEntry(any(JournalEntry.class));
+    doNothing().when(mMockJournalWriter).write(any(JournalEntry.class));
 
     // Flushes should succeed.
     for (int i = 1; i <= entries; i++) {
       mAsyncJournalWriter.flush(i);
     }
     if (batchingEnabled) {
-      Mockito.verify(mMockJournalWriter, atLeastOnce()).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, atLeastOnce()).flush();
     } else {
-      Mockito.verify(mMockJournalWriter, times(entries)).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, times(entries)).flush();
     }
   }
 
@@ -166,7 +160,7 @@ public class AsyncJournalWriterTest {
     }
 
     // Start failing journal flushes.
-    doThrow(new IOException("flush failed")).when(mMockJournalWriter).flushEntryStream();
+    doThrow(new IOException("flush failed")).when(mMockJournalWriter).flush();
 
     // Flushes should fail.
     for (int i = 1; i <= entries; i++) {
@@ -179,17 +173,17 @@ public class AsyncJournalWriterTest {
     }
 
     // Allow journal flushes to succeed.
-    doNothing().when(mMockJournalWriter).flushEntryStream();
+    doNothing().when(mMockJournalWriter).flush();
 
     // Flushes should succeed.
     for (int i = 1; i <= entries; i++) {
       mAsyncJournalWriter.flush(i);
     }
     if (batchingEnabled) {
-      Mockito.verify(mMockJournalWriter, atLeastOnce()).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, atLeastOnce()).flush();
     } else {
       // The first half of the calls were the failed flush calls.
-      Mockito.verify(mMockJournalWriter, times(2 * entries)).flushEntryStream();
+      Mockito.verify(mMockJournalWriter, times(2 * entries)).flush();
     }
   }
